[#18] Refactor: CustomMenuItem 클래스 선언을 별도 파일로 분리
