+++ PhotoAlbum/PhotoAlbum/Doodle/DoodleViewController.swift
@@ -84,13 +84,4 @@ extension DoodleViewController: UICollectionViewDataSource, UICollectionViewDele
     
 }
 
-class CustomMenuItem: UIMenuItem {
-    var indexPath: IndexPath?
-    
-    convenience init(title: String, action: Selector, indexPath: IndexPath? = nil) {
-        self.init(title: title, action: action)
-        self.indexPath = indexPath
-    }
-}
-
